src/requirements.txt
psycopg2-binary==2.9.3
kafka-python==2.0.2
faker==13.3.4
apache-flink==1.14.0 
numpy==1.24.0 
pytest==7.1.1
black==22.3.0
flake8==4.0.1
pyyaml==6.0
python-dotenv==0.20.0
